css/cssom-view/scrollIntoView-vertical-rl-writing-mode.html

<!DOCTYPE html>
<title>CSSOM View - scrollIntoView considers vertical-rl writing mode</title>
<meta charset="utf-8">
<link rel="author" title="Suneel Kota" href="mailto:suneel.kota@samsung.com">
<link rel="help" href="https://drafts.csswg.org/cssom-view/#dom-element-scrollintoview">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.box {
  float: left;
  width: 200px;
  height: 200px;
}
#scroller {
  overflow: hidden;
  width: 300px;
  height: 300px;
}
#container{
  writing-mode: vertical-rl;
  width: 600px;
  height: 600px;
}
#target {
  background-color: #ff0;
}
</style>
<body>
<div id="scroller">
  <div id="container">
    <!-- ROW-1 -->
    <div class="row">
      <div class="box"></div>
      <div class="box"></div>
      <div class="box"></div>
    </div>

    <!-- ROW-2 -->
    <div class="row">
      <div class="box"></div>
      <div class="box" id="target"></div>
      <div class="box"></div>
    </div>

    <!-- ROW-3 -->
    <div class="row">
      <div class="box"></div>
      <div class="box"></div>
      <div class="box"></div>
    </div>
  </div>
</div>

<script>
var target = document.getElementById("target");
var scroller = document.getElementById("scroller");

var scroller_width = scroller.offsetWidth;
var scroller_height = scroller.offsetHeight;
var box_width = target.offsetWidth;
var box_height = target.offsetHeight;

var calculatedX = [ ((2*box_width)-scroller_width), ((3*box_width - scroller_width)/2), box_width ];
var calculatedY = [ box_height, ((3*box_height - scroller_height)/2), ((2*box_height)-scroller_height) ];

// This formats dict as a string suitable as test name.
// format_value() is provided by testharness.js,
// which also preserves sign for -0.
function format_dict(dict) {
  const props = [];
  for (let prop in dict) {
    props.push(`${prop}: ${format_value(dict[prop])}`);
  }
  return `{${props.join(", ")}}`;
}

[
  [{block: "start", inline: "start"}, calculatedX[0], calculatedY[0]],
  [{block: "start", inline: "center"}, calculatedX[0], calculatedY[1]],
  [{block: "start", inline: "end"}, calculatedX[0], calculatedY[2]],
  [{block: "center", inline: "start"}, calculatedX[1], calculatedY[0]],
  [{block: "center", inline: "center"}, calculatedX[1], calculatedY[1]],
  [{block: "center", inline: "end"}, calculatedX[1], calculatedY[2]],
  [{block: "end", inline: "start"}, calculatedX[2], calculatedY[0]],
  [{block: "end", inline: "center"}, calculatedX[2], calculatedY[1]],
  [{block: "end", inline: "end"}, calculatedX[2], calculatedY[2]],
].forEach(([input, expectedX, expectedY]) => {
  test(() => {
    scroller.scrollTo(0, 0);
    target.scrollIntoView(input);
    assert_approx_equals(scroller.scrollLeft, expectedX, 0.5, "scrollX");
    assert_approx_equals(scroller.scrollTop, expectedY, 0.5, "scrollY");
  }, `scrollIntoView(${format_dict(input)})`);
})
</script>

</body>
</html>
